extension/backend/WEB_CRAWLER_SIH/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
PyMySQL==1.1.0
cryptography==41.0.7
requests==2.31.0
playwright==1.40.0
Pillow==10.1.0
opencv-python==4.8.1.78
easyocr==1.7.1
torch==2.1.1
torchvision==0.20.1
numpy==1.24.3
python-dateutil==2.8.2
python-dotenv==1.0.0
boto3==1.34.10
gunicorn==21.2.0
APScheduler==3.10.4
